bigger nums - might switch to float soon

+++ src/atom.rs
@@ -1,11 +1,11 @@
 #[derive(Debug)]
 pub struct Nucleus {
-    pub protons:  i32,
-    pub neutrons: i32,
+    pub protons:  i64,
+    pub neutrons: i64,
 }
 
 #[derive(Debug)]
 pub struct Atom {
-    pub electrons: i32,
+    pub electrons: i64,
     pub nucleus: Nucleus,
 }
